requirements.txt
APScheduler==3.6.3
async-generator==1.10
attrs==22.1.0
cachetools==4.2.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.0.0
click==8.0.0
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
exceptiongroup==1.0.0
Flask==2.2.2
fonttools==4.38.0
gunicorn==20.0.4
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.0.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.1
numpy==1.23.4
outcome==1.2.0
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
psutil==5.9.3
pycparser==2.21
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-telegram-bot==13.14
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
redis==4.3.4
requests==2.28.1
selenium==4.4.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
telegram==0.0.1
tornado==6.1
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
Werkzeug==2.2.2
wsproto==1.2.0

